exchange-backend/requirements.txt
apispec==4.4.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.10.0
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
